// src/users/interfaces/user.interfaces.ts
/** @TODO EUserStatus */
export enum EUserStatus {
  ACTIVE = 1,
  SUSPENDED = 2,
  UNCONFIRMED = 3,
}

export enum EUserGender {
  OTHER = 1,
  FEMALE = 2,
  MALE = 3,
}

export enum EUserRank {
  NEW = 1,
  BRONZE = 2,
  SILVER = 3,
  GOLS = 4,
  PLATINUM = 5,
  DIAMON = 6,
  VIP = 7,
  VIP_3 = 8,
}